+++ core/src/test/scala/integration/kafka/api/SaslSslAdminIntegrationTest.scala
@@ -167,93 +167,6 @@ class SaslSslAdminIntegrationTest extends BaseAdminIntegrationTest with SaslSetu
   }
 
   @Test
-  def testAclDelete(): Unit = {
-    client = Admin.create(createConfig)
-    ensureAcls(Set(anyAcl, acl2, fooAcl, prefixAcl))
-
-    val allTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.MATCH), AccessControlEntryFilter.ANY)
-    val allLiteralTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.LITERAL), AccessControlEntryFilter.ANY)
-    val allPrefixedTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.PREFIXED), AccessControlEntryFilter.ANY)
-
-    // Delete only ACLs on literal 'mytopic2' topic
-    var deleted = client.deleteAcls(List(acl2.toFilter).asJava).all().get().asScala.toSet
-    assertEquals(Set(acl2), deleted)
-    assertEquals(Set(anyAcl, fooAcl, prefixAcl), getAcls(allTopicAcls))
-
-    ensureAcls(deleted)
-
-    // Delete only ACLs on literal '*' topic
-    deleted = client.deleteAcls(List(anyAcl.toFilter).asJava).all().get().asScala.toSet
-    assertEquals(Set(anyAcl), deleted)
-    assertEquals(Set(acl2, fooAcl, prefixAcl), getAcls(allTopicAcls))
-
-    ensureAcls(deleted)
-
-    // Delete only ACLs on specific prefixed 'mytopic' topics:
-    deleted = client.deleteAcls(List(prefixAcl.toFilter).asJava).all().get().asScala.toSet
-    assertEquals(Set(prefixAcl), deleted)
-    assertEquals(Set(anyAcl, acl2, fooAcl), getAcls(allTopicAcls))
-
-    ensureAcls(deleted)
-
-    // Delete all literal ACLs:
-    deleted = client.deleteAcls(List(allLiteralTopicAcls).asJava).all().get().asScala.toSet
-    assertEquals(Set(anyAcl, acl2, fooAcl), deleted)
-    assertEquals(Set(prefixAcl), getAcls(allTopicAcls))
-
-    ensureAcls(deleted)
-
-    // Delete all prefixed ACLs:
-    deleted = client.deleteAcls(List(allPrefixedTopicAcls).asJava).all().get().asScala.toSet
-    assertEquals(Set(prefixAcl), deleted)
-    assertEquals(Set(anyAcl, acl2, fooAcl), getAcls(allTopicAcls))
-
-    ensureAcls(deleted)
-
-    // Delete all topic ACLs:
-    deleted = client.deleteAcls(List(allTopicAcls).asJava).all().get().asScala.toSet
-    assertEquals(Set(), getAcls(allTopicAcls))
-  }
-
-  //noinspection ScalaDeprecation - test explicitly covers clients using legacy / deprecated constructors
-  @Test
-  def testLegacyAclOpsNeverAffectOrReturnPrefixed(): Unit = {
-    client = Admin.create(createConfig)
-    ensureAcls(Set(anyAcl, acl2, fooAcl, prefixAcl))  // <-- prefixed exists, but should never be returned.
-
-    val allTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.MATCH), AccessControlEntryFilter.ANY)
-    val legacyAllTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.LITERAL), AccessControlEntryFilter.ANY)
-    val legacyMyTopic2Acls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, "mytopic2", PatternType.LITERAL), AccessControlEntryFilter.ANY)
-    val legacyAnyTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, "*", PatternType.LITERAL), AccessControlEntryFilter.ANY)
-    val legacyFooTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, "foobar", PatternType.LITERAL), AccessControlEntryFilter.ANY)
-
-    assertEquals(Set(anyAcl, acl2, fooAcl), getAcls(legacyAllTopicAcls))
-    assertEquals(Set(acl2), getAcls(legacyMyTopic2Acls))
-    assertEquals(Set(anyAcl), getAcls(legacyAnyTopicAcls))
-    assertEquals(Set(fooAcl), getAcls(legacyFooTopicAcls))
-
-    // Delete only (legacy) ACLs on 'mytopic2' topic
-    var deleted = client.deleteAcls(List(legacyMyTopic2Acls).asJava).all().get().asScala.toSet
-    assertEquals(Set(acl2), deleted)
-    assertEquals(Set(anyAcl, fooAcl, prefixAcl), getAcls(allTopicAcls))
-
-    ensureAcls(deleted)
-
-    // Delete only (legacy) ACLs on '*' topic
-    deleted = client.deleteAcls(List(legacyAnyTopicAcls).asJava).all().get().asScala.toSet
-    assertEquals(Set(anyAcl), deleted)
-    assertEquals(Set(acl2, fooAcl, prefixAcl), getAcls(allTopicAcls))
-
-    ensureAcls(deleted)
-
-    // Delete all (legacy) topic ACLs:
-    deleted = client.deleteAcls(List(legacyAllTopicAcls).asJava).all().get().asScala.toSet
-    assertEquals(Set(anyAcl, acl2, fooAcl), deleted)
-    assertEquals(Set(), getAcls(legacyAllTopicAcls))
-    assertEquals(Set(prefixAcl), getAcls(allTopicAcls))
-  }
-
-  @Test
   def testAttemptToCreateInvalidAcls(): Unit = {
     client = Admin.create(createConfig)
     val clusterAcl = new AclBinding(new ResourcePattern(ResourceType.CLUSTER, "foobar", PatternType.LITERAL),
