+++ core/src/test/scala/integration/kafka/api/SaslSslAdminIntegrationTest2.scala
@@ -13,20 +13,22 @@
 package kafka.api
 
 import java.io.File
+import java.util
 import kafka.security.authorizer.AclAuthorizer
 import kafka.security.authorizer.AclEntry.{WildcardHost, WildcardPrincipalString}
 import kafka.server.KafkaConfig
 import kafka.utils.{CoreUtils, JaasTestUtils, TestUtils}
+import org.apache.kafka.clients.admin.Admin
 import org.apache.kafka.common.acl._
 import org.apache.kafka.common.acl.AclOperation.{ALL, ALTER, ALTER_CONFIGS, CLUSTER_ACTION, CREATE, DELETE}
 import org.apache.kafka.common.acl.AclPermissionType.ALLOW
 import org.apache.kafka.common.resource.PatternType.LITERAL
 import org.apache.kafka.common.resource.ResourceType.{GROUP, TOPIC}
-import org.apache.kafka.common.resource.{PatternType, Resource, ResourcePattern, ResourceType}
+import org.apache.kafka.common.resource.{PatternType, Resource, ResourcePattern, ResourcePatternFilter, ResourceType}
 import org.apache.kafka.common.security.auth.{KafkaPrincipal, SecurityProtocol}
 import org.apache.kafka.server.authorizer.Authorizer
 import org.junit.jupiter.api.Assertions._
-import org.junit.jupiter.api.{AfterEach, BeforeEach}
+import org.junit.jupiter.api.{AfterEach, BeforeEach, Test}
 
 import java.util.Collections
 import scala.jdk.CollectionConverters._
@@ -67,12 +69,123 @@ class SaslSslAdminIntegrationTest2 extends BaseAdminIntegrationTest2 with SaslSe
     closeSasl()
   }
 
+  val anyAcl = new AclBinding(new ResourcePattern(ResourceType.TOPIC, "*", PatternType.LITERAL),
+    new AccessControlEntry("User:*", "*", AclOperation.ALL, AclPermissionType.ALLOW))
   val acl2 = new AclBinding(new ResourcePattern(ResourceType.TOPIC, "mytopic2", PatternType.LITERAL),
     new AccessControlEntry("User:ANONYMOUS", "*", AclOperation.WRITE, AclPermissionType.ALLOW))
+  val fooAcl = new AclBinding(new ResourcePattern(ResourceType.TOPIC, "foobar", PatternType.LITERAL),
+    new AccessControlEntry("User:ANONYMOUS", "*", AclOperation.READ, AclPermissionType.ALLOW))
+  val prefixAcl = new AclBinding(new ResourcePattern(ResourceType.TOPIC, "mytopic", PatternType.PREFIXED),
+    new AccessControlEntry("User:ANONYMOUS", "*", AclOperation.READ, AclPermissionType.ALLOW))
 
   override def configuredClusterPermissions: Set[AclOperation] = {
     Set(AclOperation.ALTER, AclOperation.CREATE, AclOperation.CLUSTER_ACTION, AclOperation.ALTER_CONFIGS,
       AclOperation.DESCRIBE, AclOperation.DESCRIBE_CONFIGS)
+  }
+
+  @Test
+  def testAclDelete(): Unit = {
+    client = Admin.create(createConfig)
+    ensureAcls(Set(anyAcl, acl2, fooAcl, prefixAcl))
+
+    val allTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.MATCH), AccessControlEntryFilter.ANY)
+    val allLiteralTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.LITERAL), AccessControlEntryFilter.ANY)
+    val allPrefixedTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.PREFIXED), AccessControlEntryFilter.ANY)
+
+    // Delete only ACLs on literal 'mytopic2' topic
+    var deleted = client.deleteAcls(List(acl2.toFilter).asJava).all().get().asScala.toSet
+    assertEquals(Set(acl2), deleted)
+    assertEquals(Set(anyAcl, fooAcl, prefixAcl), getAcls(allTopicAcls))
+
+    ensureAcls(deleted)
+
+    // Delete only ACLs on literal '*' topic
+    deleted = client.deleteAcls(List(anyAcl.toFilter).asJava).all().get().asScala.toSet
+    assertEquals(Set(anyAcl), deleted)
+    assertEquals(Set(acl2, fooAcl, prefixAcl), getAcls(allTopicAcls))
+
+    ensureAcls(deleted)
+
+    // Delete only ACLs on specific prefixed 'mytopic' topics:
+    deleted = client.deleteAcls(List(prefixAcl.toFilter).asJava).all().get().asScala.toSet
+    assertEquals(Set(prefixAcl), deleted)
+    assertEquals(Set(anyAcl, acl2, fooAcl), getAcls(allTopicAcls))
+
+    ensureAcls(deleted)
+
+    // Delete all literal ACLs:
+    deleted = client.deleteAcls(List(allLiteralTopicAcls).asJava).all().get().asScala.toSet
+    assertEquals(Set(anyAcl, acl2, fooAcl), deleted)
+    assertEquals(Set(prefixAcl), getAcls(allTopicAcls))
+
+    ensureAcls(deleted)
+
+    // Delete all prefixed ACLs:
+    deleted = client.deleteAcls(List(allPrefixedTopicAcls).asJava).all().get().asScala.toSet
+    assertEquals(Set(prefixAcl), deleted)
+    assertEquals(Set(anyAcl, acl2, fooAcl), getAcls(allTopicAcls))
+
+    ensureAcls(deleted)
+
+    // Delete all topic ACLs:
+    deleted = client.deleteAcls(List(allTopicAcls).asJava).all().get().asScala.toSet
+    assertEquals(Set(), getAcls(allTopicAcls))
+  }
+
+  //noinspection ScalaDeprecation - test explicitly covers clients using legacy / deprecated constructors
+  @Test
+  def testLegacyAclOpsNeverAffectOrReturnPrefixed(): Unit = {
+    client = Admin.create(createConfig)
+    ensureAcls(Set(anyAcl, acl2, fooAcl, prefixAcl))  // <-- prefixed exists, but should never be returned.
+
+    val allTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.MATCH), AccessControlEntryFilter.ANY)
+    val legacyAllTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, null, PatternType.LITERAL), AccessControlEntryFilter.ANY)
+    val legacyMyTopic2Acls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, "mytopic2", PatternType.LITERAL), AccessControlEntryFilter.ANY)
+    val legacyAnyTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, "*", PatternType.LITERAL), AccessControlEntryFilter.ANY)
+    val legacyFooTopicAcls = new AclBindingFilter(new ResourcePatternFilter(ResourceType.TOPIC, "foobar", PatternType.LITERAL), AccessControlEntryFilter.ANY)
+
+    assertEquals(Set(anyAcl, acl2, fooAcl), getAcls(legacyAllTopicAcls))
+    assertEquals(Set(acl2), getAcls(legacyMyTopic2Acls))
+    assertEquals(Set(anyAcl), getAcls(legacyAnyTopicAcls))
+    assertEquals(Set(fooAcl), getAcls(legacyFooTopicAcls))
+
+    // Delete only (legacy) ACLs on 'mytopic2' topic
+    var deleted = client.deleteAcls(List(legacyMyTopic2Acls).asJava).all().get().asScala.toSet
+    assertEquals(Set(acl2), deleted)
+    assertEquals(Set(anyAcl, fooAcl, prefixAcl), getAcls(allTopicAcls))
+
+    ensureAcls(deleted)
+
+    // Delete only (legacy) ACLs on '*' topic
+    deleted = client.deleteAcls(List(legacyAnyTopicAcls).asJava).all().get().asScala.toSet
+    assertEquals(Set(anyAcl), deleted)
+    assertEquals(Set(acl2, fooAcl, prefixAcl), getAcls(allTopicAcls))
+
+    ensureAcls(deleted)
+
+    // Delete all (legacy) topic ACLs:
+    deleted = client.deleteAcls(List(legacyAllTopicAcls).asJava).all().get().asScala.toSet
+    assertEquals(Set(anyAcl, acl2, fooAcl), deleted)
+    assertEquals(Set(), getAcls(legacyAllTopicAcls))
+    assertEquals(Set(prefixAcl), getAcls(allTopicAcls))
+  }
+
+  private def waitForDescribeAcls(client: Admin, filter: AclBindingFilter, acls: Set[AclBinding]): Unit = {
+    var lastResults: util.Collection[AclBinding] = null
+    TestUtils.waitUntilTrue(() => {
+      lastResults = client.describeAcls(filter).values.get()
+      acls == lastResults.asScala.toSet
+    }, s"timed out waiting for ACLs $acls.\nActual $lastResults")
+  }
+
+  private def ensureAcls(bindings: Set[AclBinding]): Unit = {
+    client.createAcls(bindings.asJava).all().get()
+
+    bindings.foreach(binding => waitForDescribeAcls(client, binding.toFilter, Set(binding)))
+  }
+
+  private def getAcls(allTopicAcls: AclBindingFilter) = {
+    client.describeAcls(allTopicAcls).values.get().asScala.toSet
   }
 
   class AclAuthorizationAdmin(authorizerClass: Class[_ <: AclAuthorizer], authorizerForInitClass: Class[_ <: AclAuthorizer]) {
